gen-ananlytics-backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
openai==1.1.1
python-dotenv==1.0.0
httpx==0.25.0
